+++ projects/ng-wormhole/src/lib/ng-wormhole.directive.spec.ts
@@ -12,19 +12,17 @@ function removeHTMLComments(str: string) {
   selector: 'ngw-test',
   template: `
     <div id="target-element"></div>
-    <ng-wormhole to="#target-element">
-      <h1>welcome</h1>
-    </ng-wormhole>
+    <div id="wormhole"><h1 *ngWormhole="'#target-element'">welcome</h1></div>
   `
 })
 class TestComponent {}
 
-describe('NgWormholeComponent Acceptance', () => {
+describe('NgWormholeDirective Acceptance', () => {
   let fixture: ComponentFixture<TestComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [TestComponent, NgWormholeComponent, NgWormholeDirective]
+      declarations: [TestComponent, NgWormholeDirective]
     }).compileComponents();
   }));
 
@@ -40,8 +38,8 @@ describe('NgWormholeComponent Acceptance', () => {
     expect(targetContent).toBe('<h1>welcome</h1>');
   });
 
-  it('should be empty', () => {
-    let wormholeContent = fixture.debugElement.query(By.css('ng-wormhole'))
+  it('should not be rendered outside target', () => {
+    let wormholeContent = fixture.debugElement.query(By.css('#wormhole'))
       .nativeElement.innerHTML;
     wormholeContent = removeHTMLComments(wormholeContent);
     expect(wormholeContent).toBe('');
@@ -52,19 +50,17 @@ describe('NgWormholeComponent Acceptance', () => {
   selector: 'ngw-test',
   template: `
     <div id="target-element"></div>
-    <ng-wormhole to="#target-element" [renderInPlace]="true">
-      <h1>welcome</h1>
-    </ng-wormhole>
+    <div id="wormhole"><h1 *ngWormhole="'#target-element'; renderInPlace: true">welcome</h1></div>
   `
 })
 class Test2Component {}
 
-describe('NgWormholeComponent Acceptance renderInPlace', () => {
+describe('NgWormholeDirective Acceptance renderInPlace', () => {
   let fixture: ComponentFixture<Test2Component>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [Test2Component, NgWormholeComponent, NgWormholeDirective]
+      declarations: [Test2Component, NgWormholeDirective]
     }).compileComponents();
   }));
 
@@ -81,9 +77,47 @@ describe('NgWormholeComponent Acceptance renderInPlace', () => {
   });
 
   it('should not be empty', () => {
-    let wormholeContent = fixture.debugElement.query(By.css('ng-wormhole'))
+    let wormholeContent = fixture.debugElement.query(By.css('#wormhole'))
       .nativeElement.innerHTML;
     wormholeContent = removeHTMLComments(wormholeContent);
     expect(wormholeContent).toBe('<h1>welcome</h1>');
+  });
+});
+
+@Component({
+  selector: 'ngw-test',
+  template: `
+    <div #targetElement id="target-element"></div>
+    <div id="wormhole"><h1 *ngWormhole="targetElement">welcome</h1></div>
+  `
+})
+class Test3Component {}
+
+describe('NgWormholeDirective Acceptance elementRef', () => {
+  let fixture: ComponentFixture<Test3Component>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [Test3Component, NgWormholeDirective]
+    }).compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(Test3Component);
+    fixture.detectChanges();
+  });
+
+  it('should render into target', () => {
+    let targetContent = fixture.debugElement.query(By.css('#target-element'))
+      .nativeElement.innerHTML;
+    targetContent = removeHTMLComments(targetContent);
+    expect(targetContent).toBe('<h1>welcome</h1>');
+  });
+
+  it('should be empty', () => {
+    let wormholeContent = fixture.debugElement.query(By.css('#wormhole'))
+      .nativeElement.innerHTML;
+    wormholeContent = removeHTMLComments(wormholeContent);
+    expect(wormholeContent).toBe('');
   });
 });