+++ projects/ng-wormhole/src/lib/ng-wormhole.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { NgWormholeComponent } from './ng-wormhole.component';
+import { NgWormholeDirective } from './ng-wormhole.directive';
 
 @NgModule({
-  imports: [
-  ],
-  declarations: [NgWormholeComponent],
-  exports: [NgWormholeComponent]
+  imports: [],
+  declarations: [NgWormholeComponent, NgWormholeDirective],
+  exports: [NgWormholeComponent, NgWormholeDirective]
 })
-export class NgWormholeModule { }
+export class NgWormholeModule {}