+++ projects/ng-wormhole/src/public_api.ts
@@ -3,4 +3,5 @@
  */
 
 export * from './lib/ng-wormhole.component';
+export * from './lib/ng-wormhole.directive';
 export * from './lib/ng-wormhole.module';